+++ native/kotlin/api/kotlin/src/main/kotlin/rs/wordpress/api/kotlin/WpComApiClient.kt
@@ -55,12 +55,15 @@ class WpComApiClient(
     //
     // It'll also help make sure any breaking changes to the API will end up as a compiler error.
     suspend fun <T> request(
+        logErrors: Boolean = true,
         executeRequest: suspend (UniffiWpComApiClient) -> T
     ): WpRequestResult<T> = withContext(dispatcher) {
-        try {
+        val result = try {
             WpRequestResult.Success(response = executeRequest(requestBuilder))
         } catch (exception: WpApiException) {
-            mapWpApiExceptionToWpRequestResult(exception)
+            mapWpApiExceptionToWpRequestResult<T>(exception)
         }
+        result.logErrorIfNeeded(logErrors)
+        result
     }
 }